tests/requirements.txt
requests==2.23.0	
PyJWT==1.7.1	
cryptography==2.9.2	
unittest-xml-reporting==2.4.0	
json-logging==1.2.0